fix benign bug in ringfix (no effect observed in 1M mols)
a minus sign was missing in the the logic to choose the best
conformer. Seems like the magnitude of the tested value, which is
the difference in axial likeliness of two conformers, is much
larger than the cutoff value (0.1), so we are never in a situation
where delta_axial_likeliness is between -0.1 and +0.1.